chore: Rename empty state components
